perf!: true logins (WIP)
I've been leaning on the database up until now for resolving user
logins and permissions. However this will severely limit the performance
of the application.

Thus it is time to implement a true Winvoice login system. For now,
users still have to be created by a DB admin, and there is no way to
change a password without their intervention. This can change with time,
however.

* `argon2` will be used to hash user passwords in the database.
* `casbin` will be used to resolve user permissions, and those
  permissions can be reloaded without restarting the server using
  `watchman_client`.
* `axum-sessions` and `axum-login` will be used to manage user sessions.
  I ain't doin all that myself.