fix: ensure password set date is updated when password is changed
This requires changing the `password_expires` date to be a
`password_set` field. Consider the following example: a `User` changes
their password, while being assigned to a `Role` which does not enforce
password rotation. Then, later they are promoted to a `Role` which
*does* require password rotation. The `User` would have no expiry date,
no way to track the original date of password set.

This change also allows the `User`'s `Role` to be
adjusted without the password expiry date remaining static.